New feature for percentage prefix_length

+++ src/encoding/models.py
@@ -62,7 +62,7 @@ class Encoding(CommonModel):
     add_resources_used = models.BooleanField(default=False)
     add_new_traces = models.BooleanField(default=False)
     features = JSONField(default=dict)
-    prefix_length = models.PositiveIntegerField()
+    prefix_length = models.FloatField()
     padding = models.BooleanField(default=False)
     task_generation_type = models.CharField(choices=TASK_GENERATION_TYPE_MAPPINGS, default='only_this', max_length=max(len(el[1]) for el in TASK_GENERATION_TYPE_MAPPINGS)+1)
